refactor: Remove AST converter class
