+++ src/eslint-adapter.test.ts
@@ -1,7 +1,6 @@
 import path from "path";
 import ts from "typescript";
 import { mark, Frets } from "fretted-strings";
-import { AstConverter } from "./ast-converter";
 import { ESLintAdapter } from "./eslint-adapter";
 import { ConfigProvider } from "./eslint-config-provider";
 import { Legacy } from "@eslint/eslintrc";
@@ -39,7 +38,6 @@ describe("ESLintAdapter", () => {
         semi: 2,
       };
       const adapter = new ESLintAdapter({
-        converter: new AstConverter(),
         getSourceFile: () => ts.createSourceFile("main.ts", "const x = 1", ts.ScriptTarget.ESNext, true),
         configProvider,
         logger: () => {},
@@ -65,7 +63,6 @@ describe("ESLintAdapter", () => {
         frets,
       );
       const adapter = new ESLintAdapter({
-        converter: new AstConverter(),
         getSourceFile: () => ts.createSourceFile("main.ts", content, ts.ScriptTarget.ESNext, true),
         configProvider,
         logger: () => {},
@@ -91,7 +88,6 @@ describe("ESLintAdapter", () => {
         frets,
       );
       const adapter = new ESLintAdapter({
-        converter: new AstConverter(),
         getSourceFile: () => ts.createSourceFile("main.ts", content, ts.ScriptTarget.ESNext, true),
         configProvider,
         logger: () => {},
@@ -123,7 +119,6 @@ describe("ESLintAdapter", () => {
         frets,
       );
       const adapter = new ESLintAdapter({
-        converter: new AstConverter(),
         getSourceFile: () => ts.createSourceFile("main.ts", content, ts.ScriptTarget.ESNext, true),
         configProvider,
         logger: () => {},
